Add files via upload
It won’t let me put it on the GitHub uncompressed for some reason but here are all the screenshots in a word document. When he said that it doesn’t matter right now which OS it works on I went ahead and took screenshots so it’ll look pretty in the README :) 

What do we need in that file?

Hope they’re okay!